+++ desktop/cmp/input/Select.js
@@ -303,7 +303,6 @@ class Model extends HoistInputModel {
             if (action === 'input-change') {
                 this.inputValue = value;
                 this.inputValueChangedSinceSelect = true;
-                if (!value) this.noteValueChange(null);
             } else if (action === 'input-blur') {
                 this.inputValue = null;
                 this.inputValueChangedSinceSelect = false;
@@ -652,6 +651,7 @@ const cmp = hoistCmp.factory(
         if (model.manageInputValue) {
             rsProps.inputValue = model.inputValue || '';
             rsProps.onInputChange = model.onInputChange;
+            rsProps.controlShouldRenderValue = !model.hasFocus;
             rsProps.onMenuOpen = () => {
                 wait().then(()=> {
                     const selectedEl = document.getElementsByClassName('xh-select__option--is-selected')[0];