+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 * Added `GridModel.restoreDefaultsFn`, an optional function allowing apps to extend
   `GridModel.restoreDefaultsAsync()` to additionally reset app-specific state.
 
+### 🐞 Bug Fixes
+
+* Fix column auto-sizing when `headerName` is/returns an element
+
 ## v47.0.1 - 2022-03-06
 
 [Commit Log](https://github.com/xh/hoist-react/compare/v47.0.0...v47.0.1)