+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## v48.0.0-SNAPSHOT- unreleased
 
+### 🎁 New Features
+
+* Configs passed to `XH.message()` and variants support a new `cancelAlign` property. Customize
+  to "left" to render Cancel and Confirm actions separated by a filler.
+
 ### 🐞 Bug Fixes
 
 * Fix column auto-sizing when `headerName` is/returns an element