+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 ### 🐞 Bug Fixes
 
 * Fix column auto-sizing when `headerName` is/returns an element
+* Fixed an issue where `Select` inputs would commit `null` whilst clearing the text input.
 
 ### 📚 Libraries